requirements.txt
PyPDF2>=3.0.0
pikepdf>=8.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0
reportlab>=4.0.0
pdfplumber>=0.10.0
img2pdf>=0.5.0
PyYAML>=6.0.0
tqdm>=4.66.0
Flask>=3.0.0
streamlit>=1.40.0
